Add register-server
The register server is a new service that a machine joining the cluster
needs to reach in order to get an ID. This ID will be then used to
recognize the node for the secrets and other infrastructure components
that needs to identify the node.

The register servers for new registered will create a corresponding
machine object with the released id and IP of the request. For new
registration request coming from the IP, the old machine object will be
replaced by a new one.

Signed-off-by: Alice Frosi <afrosi@redhat.com>
Co-authored-by: Jakob Naucke <jnaucke@redhat.com>