Include Azure SNP vTPM in the policies
In the resource policy we cannot affirm that the cpu is tpm. Let's fix
later. Now, the operator only compares PCRs register. Unfortunately, the
PCR 7 depends on the secure boot variables which are different on Azure.
Hence, the calculation done by the compute library still doesn't take
this into account.