Check cluster is installed in reconciliation
PR #80 deliberately made cluster updates impossible and thus stopped
checking for new cluster versions. However, the reason for the check
persisted: Because the status is updated twice, reconciliation will
oscillate between "installing" and "installed" forever at a very high
rate.

Because updates are no longer possible, we can now perform a more
stable check: If the cluster is set to installed and is also not being
deleted, consider reconciled.

Signed-off-by: Jakob Naucke <jnaucke@redhat.com>